Respect DPI
pdf.py is just a quick hack and I don't want to extend it to
introspecting the JBIG2 bitstreams. So I just hardcode the DPI value
here in the hope that the size of the resulting pages will be a little
saner for most people and that it's obvious what to change now.